requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.10
django-cors-headers==3.11.0
django-model-utils==4.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
Unipath==1.1
